Change atomInfo back from 64 to 32 bits
There is no need for more than 32 bits of information.

+++ api/nblib/gmxcalculatorcpu.cpp
@@ -71,7 +71,7 @@ public:
     CpuImpl(gmx::ArrayRef<int>     particleTypeIdOfAllParticles,
             gmx::ArrayRef<real>    nonBondedParams,
             gmx::ArrayRef<real>    charges,
-            gmx::ArrayRef<int64_t> particleInteractionFlags,
+            gmx::ArrayRef<int32_t> particleInteractionFlags,
             gmx::ArrayRef<int>     exclusionRanges,
             gmx::ArrayRef<int>     exclusionElements,
             const NBKernelOptions& options);
@@ -108,7 +108,7 @@ private:
 GmxNBForceCalculatorCpu::CpuImpl::CpuImpl(gmx::ArrayRef<int>     particleTypeIdOfAllParticles,
                                           gmx::ArrayRef<real>    nonBondedParams,
                                           gmx::ArrayRef<real>    charges,
-                                          gmx::ArrayRef<int64_t> particleInteractionFlags,
+                                          gmx::ArrayRef<int32_t> particleInteractionFlags,
                                           gmx::ArrayRef<int>     exclusionRanges,
                                           gmx::ArrayRef<int>     exclusionElements,
                                           const NBKernelOptions& options) :
@@ -270,7 +270,7 @@ void GmxNBForceCalculatorCpu::CpuImpl::compute(gmx::ArrayRef<const gmx::RVec> co
 GmxNBForceCalculatorCpu::GmxNBForceCalculatorCpu(gmx::ArrayRef<int>  particleTypeIdOfAllParticles,
                                                  gmx::ArrayRef<real> nonBondedParams,
                                                  gmx::ArrayRef<real> charges,
-                                                 gmx::ArrayRef<int64_t> particleInteractionFlags,
+                                                 gmx::ArrayRef<int32_t> particleInteractionFlags,
                                                  gmx::ArrayRef<int>     exclusionRanges,
                                                  gmx::ArrayRef<int>     exclusionElements,
                                                  const NBKernelOptions& options)
@@ -330,7 +330,7 @@ std::unique_ptr<GmxNBForceCalculatorCpu> setupGmxForceCalculatorCpu(const Topolo
     std::vector<real> nonBondedParameters = createNonBondedParameters(
             topology.getParticleTypes(), topology.getNonBondedInteractionMap());
 
-    std::vector<int64_t> particleInteractionFlags = createParticleInfoAllVdw(topology.numParticles());
+    std::vector<int32_t> particleInteractionFlags = createParticleInfoAllVdw(topology.numParticles());
 
     return std::make_unique<GmxNBForceCalculatorCpu>(topology.getParticleTypeIdOfAllParticles(),
                                                      nonBondedParameters,
