+++ api/nblib/systemdescription.h
@@ -73,7 +73,7 @@ struct SystemDescription
     SystemDescription(gmx::ArrayRef<int>     particleTypeIdOfAllParticles,
                       gmx::ArrayRef<real>    nonBondedParams,
                       gmx::ArrayRef<real>    charges,
-                      gmx::ArrayRef<int64_t> particleInteractionFlags)
+                      gmx::ArrayRef<int32_t> particleInteractionFlags)
     {
         std::array inputSizes{ particleTypeIdOfAllParticles.size(),
                                charges.size(),
@@ -99,7 +99,7 @@ struct SystemDescription
         nonBondedParams_ = std::vector<real>(nonBondedParams.begin(), nonBondedParams.end());
         charges_         = std::vector<real>(charges.begin(), charges.end());
         particleInfo_ =
-                std::vector<int64_t>(particleInteractionFlags.begin(), particleInteractionFlags.end());
+                std::vector<int>(particleInteractionFlags.begin(), particleInteractionFlags.end());
     }
 
     //! number of particles
@@ -118,7 +118,7 @@ struct SystemDescription
     std::vector<real> charges_;
 
     //! flag for each particle to set LJ and Q interactions
-    std::vector<int64_t> particleInfo_;
+    std::vector<int32_t> particleInfo_;
 
     //! Legacy matrix for box
     Box box_{ 0 };
