+++ src/gromacs/domdec/domdec_constraints.h
@@ -91,7 +91,7 @@ void dd_clear_local_constraint_indices(gmx_domdec_t* dd);
 int dd_make_local_constraints(struct gmx_domdec_t*           dd,
                               int                            at_start,
                               const struct gmx_mtop_t&       mtop,
-                              gmx::ArrayRef<const int64_t>   atomInfo,
+                              gmx::ArrayRef<const int32_t>   atomInfo,
                               gmx::Constraints*              constr,
                               int                            nrec,
                               gmx::ArrayRef<InteractionList> il_local);
