+++ src/gromacs/nbnxm/nbnxm.cpp
@@ -62,7 +62,7 @@ void nonbonded_verlet_t::putAtomsOnGrid(const matrix                   box,
                                         const gmx::UpdateGroupsCog*    updateGroupsCog,
                                         gmx::Range<int>                atomRange,
                                         real                           atomDensity,
-                                        gmx::ArrayRef<const int64_t>   atomInfo,
+                                        gmx::ArrayRef<const int32_t>   atomInfo,
                                         gmx::ArrayRef<const gmx::RVec> x,
                                         int                            numAtomsMoved,
                                         const int*                     move)
@@ -84,7 +84,7 @@ void nonbonded_verlet_t::putAtomsOnGrid(const matrix                   box,
 /* Calls nbnxn_put_on_grid for all non-local domains */
 void nbnxn_put_on_grid_nonlocal(nonbonded_verlet_t*              nbv,
                                 const struct gmx_domdec_zones_t* zones,
-                                gmx::ArrayRef<const int64_t>     atomInfo,
+                                gmx::ArrayRef<const int32_t>     atomInfo,
                                 gmx::ArrayRef<const gmx::RVec>   x)
 {
     for (int zone = 1; zone < zones->n; zone++)
@@ -137,7 +137,7 @@ void nonbonded_verlet_t::setLocalAtomOrder() const
 
 void nonbonded_verlet_t::setAtomProperties(gmx::ArrayRef<const int>     atomTypes,
                                            gmx::ArrayRef<const real>    atomCharges,
-                                           gmx::ArrayRef<const int64_t> atomInfo) const
+                                           gmx::ArrayRef<const int32_t> atomInfo) const
 {
     nbnxn_atomdata_set(nbat_.get(), pairSearch_->gridSet(), atomTypes, atomCharges, atomInfo);
 }
